rename globals.* package.*
