+++ src/NetPanzer/Core/main.cpp
@@ -18,7 +18,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
 
 #include "Scripts/ScriptManager.hpp"
-#include "globals.hpp"
+#include "package.hpp"
 #include "lua/lua.hpp"
 
 #ifdef WIN32
