+++ src/NetPanzer/Interfaces/InfoSocket.cpp
@@ -16,7 +16,7 @@ along with this program; if not, write to the Free Software
 Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 */
 
-#include "globals.hpp"
+#include "package.hpp"
 #include "InfoSocket.hpp"
 #include "Util/StringTokenizer.hpp"
