+++ src/NetPanzer/Views/MainMenu/MenuTemplateView.cpp
@@ -19,7 +19,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
 #include <vector>
 
-#include "globals.hpp"
+#include "package.hpp"
 #include "Views/MainMenu/MenuTemplateView.hpp"
 #include "2D/Palette.hpp"
 #include "Views/Components/Desktop.hpp"
